feat: add full component mode
