+++ src/kanban/src/Utils/toggleMenu.js
@@ -1,19 +1,9 @@
-import { HeaderDropDownMenu } from "../components/HeaderDropDownMenu/HeaderDropDownMenu";
+import { headerDropDownMenu, menuArrow, avatar } from './rerender';
 import cssHeader from "../components/Header/header.module.css";
 
-const headerDropDownMenu = new HeaderDropDownMenu();
-const menuArrow = document.getElementsByClassName(cssHeader.menu)[0];
-const avatar = document.getElementsByClassName(cssHeader.avatar)[0];
-
-const toggleMenu = () => {
+export const toggleMenu = () => {
     menuArrow.classList.toggle(cssHeader.menuOpened);
     menuArrow.classList.toggle(cssHeader.menu);
     if (menuArrow.className === cssHeader.menuOpened) avatar.appendChild(headerDropDownMenu.element);
     if (menuArrow.className === cssHeader.menu) headerDropDownMenu.element.remove();
-};
-
-console.log(menuArrow);
-menuArrow.addEventListener('click', toggleMenu);
-
-export default toggleMenu;
-
+}
