+++ src/kanban/src/components/State/State.js
@@ -1,46 +1,21 @@
-let state = [
+const state = [
     {
         title: 'Backlog',
-        issues: [
-            {
-                id: '1',
-                name: 'Sprint bugfix11 111'
-            },
-            {
-                id: '2',
-                name: 'Sprint bugfix222 22'
-            },
-            {
-                id: '3',
-                name: 'Sprint bugfix333333 3333333 33333 33333'
-            }
-        ],
+        issues: [],
     },
     {
         title: 'Ready',
         issues: [
-            {
-                id: '2',
-                name: 'Sprint bugfix222 22'
-            }
     ],
     },
     {
         title: 'In progress',
         issues: [
-            {
-                id: '3',
-                name: 'Sprint bugfix33 333'
-            }
         ],
     },
     {
         title: 'Finished',
         issues: [
-            {
-                id: '4',
-                name: 'Sprint bugfix444 44'
-            }
         ],
     },
 ];
