+++ src/kanban/src/components/Task/Task.js
@@ -3,12 +3,9 @@ import css from "./task.module.css";
 export class Task {
     constructor(text) {
         this.element = new DomElement({
-            type: "input",
+            type: "div",
             className: css.task,
             html: `${text}`,
         }).element;
-        this.element.setAttribute('placeholder', 'Add description');
-        this.element.setAttribute('type', 'text');
-        this.element.setAttribute('value', text);
     }
 }
