+++ src/kanban/src/components/Task/task.module.css
@@ -1,9 +1,11 @@
 .task {
    display: block;
-   flex-grow: auto;
    min-height: 20px;
+   min-width: max-content;
+   max-width: 350px;
+   width: 282px;
    margin: 12px;
-   padding: 6px 3px 6px 8px;
+   padding: 8px;
    background-color: #ffffff;
    font-size: 18px;
    border-radius: 5px;
